### 2025-10-04 168c1ae
Update paper.bib
