+++ paper.bib
@@ -102,11 +102,14 @@
     journal = {Journal of Open Source Software}
  }
 
- @inproceedings{statsmodels,
-  title={statsmodels: Econometric and statistical modeling with python},
-  author={Seabold, Skipper and Perktold, Josef},
-  booktitle={9th Python in Science Conference},
-  year={2010},
+@InProceedings{ statsmodels,
+  author    = { {S}kipper {S}eabold and {J}osef {P}erktold },
+  title     = { {S}tatsmodels: {E}conometric and {S}tatistical {M}odeling with {P}ython },
+  booktitle = { {P}roceedings of the 9th {P}ython in {S}cience {C}onference },
+  pages     = { 92 - 96 },
+  year      = { 2010 },
+  editor    = { {S}t\'efan van der {W}alt and {J}arrod {M}illman },
+  doi       = { 10.25080/Majora-92bf1922-011 }
 }
 
 @article{pingouin,
