### 2025-10-04 6ea9df5
Update paper.md
