### 2025-10-04 97ed13f
added code and license
