+++ code_0.7.10/dython/nominal.py
@@ -1,0 +1,1316 @@
+import concurrent.futures as cf
+import math
+import warnings
+from collections import Counter
+from itertools import repeat
+import matplotlib.pyplot as plt
+import numpy as np
+import pandas as pd
+import scipy.cluster.hierarchy as sch
+import scipy.stats as ss
+import seaborn as sns
+from psutil import cpu_count
+from typing import (
+    Union,
+    Any,
+    List,
+    Optional,
+    Callable,
+    Tuple,
+    Dict,
+    Iterable,
+    Set,
+    Literal,
+)
+from numpy.typing import NDArray, ArrayLike
+from matplotlib.colors import Colormap
+from ._private import (
+    convert,
+    remove_incomplete_samples,
+    replace_nan_with_value,
+    plot_or_not,
+)
+from .data_utils import identify_columns_by_type
+from .typing import Number, OneDimArray, TwoDimArray
+
+
+__all__ = [
+    "associations",
+    "cluster_correlations",
+    "conditional_entropy",
+    "correlation_ratio",
+    "cramers_v",
+    "identify_nominal_columns",
+    "identify_numeric_columns",
+    "numerical_encoding",
+    "replot_last_associations",
+    "theils_u",
+]
+
+_REPLACE = "replace"
+_DROP = "drop"
+_DROP_SAMPLES = "drop_samples"
+_DROP_FEATURES = "drop_features"
+_DROP_SAMPLE_PAIRS = "drop_sample_pairs"
+_SKIP = "skip"
+_DEFAULT_REPLACE_VALUE = 0.0
+_PRECISION = 1e-13
+
+_ASSOC_PLOT_PARAMS: Dict[str, Any] = dict()
+
+_NO_OP = "no-op"
+_SINGLE_VALUE_COLUMN_OP = "single-value-column-op"
+_I_EQ_J_OP = "i-equal-j-op"
+_ASSOC_OP = "assoc-op"
+
+NomNumAssocStr = Literal["correlation_ratio"]
+NumNumAssocStr = Literal["pearson", "spearman", "kendall"]
+NomNomAssocStr = Literal["cramer", "theil"]
+
+
+def _inf_nan_str(x: Union[int, float]) -> str:
+    if np.isnan(x):
+        return "NaN"
+    elif abs(x) == np.inf:
+        return "inf"
+    else:
+        return ""
+
+
+def conditional_entropy(
+    x: Union[OneDimArray, List[str]],
+    y: Union[OneDimArray, List[str]],
+    nan_strategy: str = _REPLACE,
+    nan_replace_value: Any = _DEFAULT_REPLACE_VALUE,
+    log_base: Number = math.e,
+) -> float:
+    """
+    Calculates the conditional entropy of x given y: S(x|y)
+
+    Wikipedia: https://en.wikipedia.org/wiki/Conditional_entropy
+
+    Parameters:
+    -----------
+    x : list / NumPy ndarray / Pandas Series
+        A sequence of measurements
+    y : list / NumPy ndarray / Pandas Series
+        A sequence of measurements
+    nan_strategy : string, default = 'replace'
+        How to handle missing values: can be either 'drop' to remove samples
+        with missing values, or 'replace' to replace all missing values with
+        the nan_replace_value. Missing values are None and np.nan.
+    nan_replace_value : any, default = 0.0
+        The value used to replace missing values with. Only applicable when
+        nan_strategy is set to 'replace'.
+    log_base: float, default = e
+        specifying base for calculating entropy. Default is base e.
+
+    Returns:
+    --------
+    float
+    """
+    if nan_strategy == _REPLACE:
+        x, y = replace_nan_with_value(x, y, nan_replace_value)
[…]
