### 2025-10-04 4c32b54
Update paper.md
