### 2025-10-04 #9141 @shakedzy
@danielskatz all fixed. Please note that it seems the compiler here is different than the one in your recommended GitHub Actions workflow (which I use [here](https://github.com/shakedzy/dython/blob/joss_paper/.github/workflows/draft-pdf.yml)). It seems the one that failed forces the use of `affiliations` while the one provided isn't.

I've also added the license in root and the code in directory.
