### 2025-10-04 0579a13
Update paper.md
