### 2025-10-04 #9141 @danielskatz
@shakedzy - Please add a country to your affiliation. Please also add a References header at the end of the .md file. And is there a reason you have a 2018 date in your .md file's initial metadata? Finally, for the pandas reference, please protect the author string with another {} so that it doesn't get interpreted as a first and last name.
