### 2025-10-05 0e29d20
Update paper.bib
