+++ paper.bib
@@ -22,7 +22,7 @@
 }
 
 @software{pandas,
-  author       = {The pandas development team},
+  author       = {{The pandas development team}},
   title        = {pandas-dev/pandas: Pandas},
   month        = sep,
   year         = 2025,
