### 2025-10-05 b146563
Update paper.md
