### 2025-10-05 #9141 @danielskatz
> [@danielskatz](https://github.com/danielskatz) - all done. 2018 is the original launch date of the _software_, as I assumed this is what you refer to in the metadata (as I didn't see it anywhere on the generated PDF). Let me know if there's anything else I can do in the meantime to assist the process 🙏

Please see line 10 of your paper.md file.  This is normally the date that you submit the paper.
