+++ requirements.txt
@@ -5,4 +5,3 @@ scipy>=1.7.1
 matplotlib>=3.6.0
 scikit-learn>=0.24.2
 psutil>=5.9.1
-setuptools
