### 2025-10-05 a042aa8
update typing
