### 2025-10-06 1cfd36b
Update paper.md
