+++ paper.md
@@ -7,7 +7,7 @@ authors:
 affiliations:
   - index: 1
     name: "Independent Researcher, Israel"
-date: 4 February 2018
+date: 4 October 2025
 tags: 
     - Python
     - exploratory data analysis
