### 2025-10-06 333d72d
typing refactor
