+++ tests/test_data_utils/test_split_hist.py
@@ -1,5 +1,4 @@
-import pytest
-import matplotlib.pyplot as plt
+from matplotlib.axes._axes import Axes
 
 from dython.data_utils import split_hist
 
@@ -7,4 +6,4 @@ from dython.data_utils import split_hist
 def test_split_hist_check(iris_df):
     result = split_hist(iris_df, "sepal length (cm)", "target")
 
-    assert isinstance(result, plt.Axes)
+    assert isinstance(result, Axes)
