+++ dython/nominal.py
@@ -1,35 +1,20 @@
-import concurrent.futures as cf
+
 import math
 import warnings
-from collections import Counter
-from itertools import repeat
-import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
-import scipy.cluster.hierarchy as sch
-import scipy.stats as ss
 import seaborn as sns
+import scipy.stats as ss
+import matplotlib.pyplot as plt
+import concurrent.futures as cf
+import scipy.cluster.hierarchy as sch
+from itertools import repeat
 from psutil import cpu_count
-from typing import (
-    Union,
-    Any,
-    List,
-    Optional,
-    Callable,
-    Tuple,
-    Dict,
-    Iterable,
-    Set,
-    Literal,
-)
-from numpy.typing import NDArray, ArrayLike
+from collections import Counter
 from matplotlib.colors import Colormap
-from ._private import (
-    convert,
-    remove_incomplete_samples,
-    replace_nan_with_value,
-    plot_or_not,
-)
+from matplotlib.axes._axes import Axes
+from typing import Any, Callable, Iterable, Literal, TypedDict, cast, overload
+from ._private import convert, remove_incomplete_samples, replace_nan_with_value, plot_or_not
 from .data_utils import identify_columns_by_type
 from .typing import Number, OneDimArray, TwoDimArray
 
@@ -56,7 +41,7 @@ _SKIP = "skip"
 _DEFAULT_REPLACE_VALUE = 0.0
 _PRECISION = 1e-13
 
-_ASSOC_PLOT_PARAMS: Dict[str, Any] = dict()
+_ASSOC_PLOT_PARAMS: dict[str, Any] = dict()
 
 _NO_OP = "no-op"
 _SINGLE_VALUE_COLUMN_OP = "single-value-column-op"
@@ -68,7 +53,12 @@ NumNumAssocStr = Literal["pearson", "spearman", "kendall"]
 NomNomAssocStr = Literal["cramer", "theil"]
 
 
-def _inf_nan_str(x: Union[int, float]) -> str:
+class AssociationsResult(TypedDict):
+    corr: pd.DataFrame
+    ax: Axes | None
+
+
+def _inf_nan_str(x: Number) -> str:
     if np.isnan(x):
         return "NaN"
     elif abs(x) == np.inf:
@@ -78,8 +68,8 @@ def _inf_nan_str(x: Union[int, float]) -> str:
 
 
 def conditional_entropy(
-    x: Union[OneDimArray, List[str]],
-    y: Union[OneDimArray, List[str]],
+    x: OneDimArray,
+    y: OneDimArray,
     nan_strategy: str = _REPLACE,
     nan_replace_value: Any = _DEFAULT_REPLACE_VALUE,
     log_base: Number = math.e,
@@ -125,8 +115,8 @@ def conditional_entropy(
 
 
 def cramers_v(
-    x: Union[OneDimArray, List[str]],
-    y: Union[OneDimArray, List[str]],
+    x: OneDimArray,
+    y: OneDimArray,
     bias_correction: bool = True,
     nan_strategy: str = _REPLACE,
     nan_replace_value: Any = _DEFAULT_REPLACE_VALUE,
@@ -163,7 +153,7 @@ def cramers_v(
         x, y = replace_nan_with_value(x, y, nan_replace_value)
     elif nan_strategy == _DROP:
         x, y = remove_incomplete_samples(x, y)
-    confusion_matrix = pd.crosstab(x, y)
+    confusion_matrix = pd.crosstab(x, y)   # pyright: ignore[reportArgumentType]
[…]
