+++ dython/model_utils.py
@@ -1,5 +1,6 @@
 import numpy as np
 import matplotlib.pyplot as plt
+from matplotlib.axes._axes import Axes
 from sklearn.ensemble import RandomForestClassifier
 from sklearn.metrics import roc_curve, precision_recall_curve, auc
 from sklearn.preprocessing import LabelEncoder
@@ -13,7 +14,7 @@ _ROC_PLOT_COLORS = ["b", "g", "r", "c", "m", "y", "k", "darkorange"]
 
 
 def _display_metric_plot(
-    ax: plt.Axes,
+    ax: Axes,
     metric: str,
     naives: list[tuple[Number, Number, Number, Number, str]],
     xlim: tuple[float, float],
@@ -22,7 +23,7 @@ def _display_metric_plot(
     title: str | None,
     filename: str | None,
     plot: bool,
-) -> plt.Axes:
+) -> Axes:
     for n in naives:
         ax.plot([n[0], n[1]], [n[2], n[3]], color=n[4], lw=1, linestyle="--")
     ax.set_xlim(left=xlim[0], right=xlim[1])
@@ -51,7 +52,7 @@ def _draw_estimated_optimal_threshold_mark(
     color: str,
     ms: int,
     fmt: str,
-    ax: plt.Axes,
+    ax: Axes,
 ) -> tuple[Number, Number, Number]:
     annotation_offset = (-0.027, 0.03)
     a = np.zeros((len(x_axis), 2))
@@ -64,17 +65,17 @@ def _draw_estimated_optimal_threshold_mark(
             lambda row: (1 - row[0]) ** 2 + (1 - row[1]) ** 2
         )  # optimal: (1,1)
     amin = np.apply_along_axis(dist, 1, a).argmin()
-    ax.plot(x_axis[amin], y_axis[amin], color=color, marker="o", ms=ms)
+    ax.plot(x_axis[amin], y_axis[amin], color=color, marker="o", ms=ms)     # pyright: ignore[reportCallIssue, reportArgumentType]
     ax.annotate(
-        "{th:{fmt}}".format(th=thresholds[amin], fmt=fmt),
-        xy=(x_axis[amin], y_axis[amin]),
+        "{th:{fmt}}".format(th=thresholds[amin], fmt=fmt),                  # pyright: ignore[reportCallIssue, reportArgumentType]
+        xy=(x_axis[amin], y_axis[amin]),                                    # pyright: ignore[reportCallIssue, reportArgumentType]
         color=color,
         xytext=(
-            x_axis[amin] + annotation_offset[0],
-            y_axis[amin] + annotation_offset[1],
+            x_axis[amin] + annotation_offset[0],                            # pyright: ignore[reportCallIssue, reportArgumentType, reportOperatorIssue]
+            y_axis[amin] + annotation_offset[1],                            # pyright: ignore[reportCallIssue, reportArgumentType, reportOperatorIssue]  
         ),
     )
-    return thresholds[amin], x_axis[amin], y_axis[amin]
+    return thresholds[amin], x_axis[amin], y_axis[amin]                     # pyright: ignore[reportCallIssue, reportArgumentType, reportReturnType]
 
 
 def _plot_macro_metric(
@@ -83,12 +84,12 @@ def _plot_macro_metric(
     n: int,
     lw: int,
     fmt: str,
-    ax: plt.Axes,
+    ax: Axes,
 ) -> None:
     all_x_axis = np.unique(np.concatenate([x_axis[i] for i in range(n)]))
     mean_y_axis = np.zeros_like(all_x_axis)
     for i in range(n):
-        mean_y_axis += np.interp(all_x_axis, x_axis[i], y_axis[i])
[…]

[…]
