+++ requirements.txt
@@ -1,7 +1,7 @@
-numpy>=1.23.0
-pandas>=1.4.2
-seaborn>=0.12.0
-scipy>=1.7.1
-matplotlib>=3.6.0
-scikit-learn>=0.24.2
-psutil>=5.9.1
+numpy>=2.3.3
+pandas>=2.3.3
+seaborn>=0.13.2
+scipy>=1.16.2
+matplotlib>=3.10.6
+scikit-learn>=1.7.2
+psutil>=7.1.0
