+++ requirements.txt
@@ -1,5 +1,5 @@
 numpy>=2.3.3
-pandas>=2.3.3
+pandas>=2.2.6
 seaborn>=0.13.2
 scipy>=1.16.2
 matplotlib>=3.10.6
