+++ requirements.txt
@@ -1,7 +1,7 @@
 numpy>=2.2.6
 pandas>=2.3.3
 seaborn>=0.13.2
-scipy>=1.16.2
+scipy>=1.15.3
 matplotlib>=3.10.6
 scikit-learn>=1.7.2
 psutil>=7.1.0
