+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.7.11
 * Fixing dependency issue ([#170](https://github.com/shakedzy/dython/issues/170))
+* Resolving multiple typing issues and warnings
 
 ## 0.7.10
 * _Dython now officially supports only Python 3.10 or above_
