### 2025-10-06 #9141 @simonprovost
> Do you mean that you're reviewing two papers at the same time, or that we would have multiple reviewers for this same paper? I think that both are allowed. I don't know if you get extra time, but I am happy to give you a bit more if you're willing to review this. I think that's up to me as the reviewer. What sounds like a reasonable amount of time for you?

Yes I am part of #8639. I believe that given both Dython and #8639 are in early stage review, at the very least 6 weeks no less. Between 6 to 8 weeks maximum I believe.

Your call, if that is all right, then consider me in :)

Cheers!
