### 2025-10-06 #9141 @RichardLitt
@sayantikabanik Thank you! I am here for any questions. https://joss.readthedocs.io/en/latest/reviewer_guidelines.html should also help. The review will start when we get one more reviewer.
