### 2025-10-09 #9174 @editorialbot
[REVIEW]: dython: A Set of Analysis Tools for Data and Variables in Python
**Submitting author:** @shakedzy (<a href="http://orcid.org/0009-0003-7968-6489">Shaked Zychlinski</a>)
**Repository:** https://github.com/shakedzy/dython
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v0.7.12
