### 2025-10-09 0d1b7bd
added tests and other md files
