+++ code_0.7.10/tests/test_nominal/test_associations.py
@@ -1,0 +1,137 @@
+import pytest
+import pandas as pd
+import numpy as np
+import scipy.stats as ss
+from sklearn import datasets
+from datetime import datetime, timedelta
+from matplotlib.axes._axes import Axes
+
+from dython.nominal import associations, correlation_ratio
+
+
+def test_return_type_check(iris_df):
+    assoc = associations(iris_df)
+
+    assert isinstance(assoc, dict), "associations should return a dict"
+    assert (
+        "corr" in assoc
+    ), 'associations should return a dict containing "corr" key'
+    assert (
+        "ax" in assoc
+    ), 'associations should return a dict containing "ax" key'
+
+    assert isinstance(
+        assoc["corr"], pd.DataFrame
+    ), 'assoc["corr"] should be a pandas DataFrame'
+    assert isinstance(
+        assoc["ax"], Axes
+    ), 'assoc["ax"] should be a matplotlib Axes'
+
+
+def test_dimension_check(iris_df):
+    corr = associations(iris_df)["corr"]
+    corr_shape = corr.shape
+    iris_shape = iris_df.shape
+
+    assert corr_shape[0] == corr_shape[1], "association matrix has wrong shape"
+    assert (
+        corr_shape[1] == iris_shape[1]
+    ), "association matrix has different shape from input data"
+
+
+def test_single_value_zero_association(iris_df):
+    SV_COL = 1
+    iris_df.iloc[:, SV_COL] = 42
+
+    corr = associations(iris_df)["corr"]
+
+    assert (
+        corr.iloc[:, SV_COL] == 0
+    ).all(), "single-value variable should have zero association value"
+    assert (
+        corr.iloc[SV_COL, :] == 0
+    ).all(), "single-value variable should have zero association value"
+
+
+def test_bad_nom_nom_assoc_parameter(iris_df):
+    with pytest.raises(ValueError, match="is not a supported"):
+        associations(iris_df, nom_nom_assoc="bad_parameter_name")    # pyright: ignore[reportArgumentType] -> should raise an error
+
+
+def test_bad_num_num_assoc_parameter(iris_df):
+    with pytest.raises(ValueError, match="is not a supported"):
+        associations(iris_df, num_num_assoc="bad_parameter_name")    # pyright: ignore[reportArgumentType] -> should raise an error
+
+
+def test_compute_only_ax_is_none(iris_df):
+    assoc = associations(iris_df, compute_only=True)
+
+    assert (
+        assoc["ax"] is None
+    ), 'associations with compute_only should return a None value for "ax" key'
+
+
+def test_mark_columns(iris_df):
+    corr = associations(iris_df, mark_columns=True)["corr"]
+
+    assert (
+        "(con)" in corr.index[0]
+    ), "first column should contain (con) mark if iris_df is used"
+
+
+def test_udf(iris_df):
+    def pr(x: pd.Series, y: pd.Series) -> float:
+        return ss.pearsonr(x, y)[0]   # pyright: ignore[reportReturnType]
+
+    corr1 = associations(
+        iris_df,
+        plot=False,
+        num_num_assoc="pearson",
+        nom_num_assoc="correlation_ratio",
+    )["corr"]
+    corr2 = associations(
+        iris_df, plot=False, num_num_assoc=pr, nom_num_assoc=correlation_ratio
+    )["corr"]
+    assert corr1.compare(
+        corr2
+    ).empty, (
+        "Computation of built-in measures of associations differs from UDFs"
+    )
+
+
+def test_datetime_data():
+    dt = datetime(2020, 12, 1)
+    end = datetime(2020, 12, 2)
+    step = timedelta(seconds=5)
+    result = []
[…]
