### 2025-10-30 235863a
Add first version release date for dython
Added release date information for the first version of dython.
