### 2025-11-04 #9174 @RichardLitt
Thank you @sepandhaghighi for your kind comments.

@shakedzy Could you respond to these suggestions?
