### 2025-11-04 6892e0a
added missing workflows to joss_paper
