+++ .github/workflows/pytest.yml
@@ -1,0 +1,51 @@
+name: Run pytest
+
+on:
+  workflow_dispatch: 
+  # push:
+  #   branches: [ "master", "shakedzy:master" ]
+  #   paths-ignore:
+  #     - "README.md"
+  #     - "CHANGELOG.md"
+  #     - "CODE_OF_CONDUCT.md"
+  #     - "CONTRIBUTING.md"
+  #     - "VERSION"
+  #     - "LICENSE"
+  #     - ".gitignore"
+  #     - "docs/*"
+  # pull_request:
+  #   types: [opened, reopened, edited, synchronize]
+  #   branches: [ "master", "shakedzy:master" ]
+  #   paths-ignore:
+  #     - "README.md"
+  #     - "CHANGELOG.md"
+  #     - "CODE_OF_CONDUCT.md"
+  #     - "CONTRIBUTING.md"
+  #     - "VERSION"
+  #     - "LICENSE"
+  #     - ".gitignore"
+  #     - "docs/*"
+
+permissions:
+  contents: read
+
+jobs:
+  build:
+    strategy:
+      matrix:
+        version: ["3.10", "3.13"]
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v4
+    - name: Set up Python ${{ matrix.version }}
+      uses: actions/setup-python@v5
+      with:
+        python-version: ${{ matrix.version }}
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install -r requirements.txt
+        pip install -r dev_requirements.txt
+        pip install .
+    - name: Test with pytest
+      run: pytest
