+++ .github/workflows/coverage_report.yml
@@ -4,7 +4,25 @@ on:
   workflow_dispatch: 
 
 jobs:
+  test:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v4
+    - name: Set up Python ${{ matrix.version }}
+      uses: actions/setup-python@v5
+      with:
+        python-version: ${{ matrix.version }}
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install -r requirements.txt
+        pip install -r dev_requirements.txt
+        pip install .
+    - name: Test with pytest
+      run: pytest --cov --cov-branch --cov-report=xml;
+
   create-report:
+    needs: test
     runs-on: ubuntu-latest
     steps:
     - name: Upload coverage reports to Codecov
