+++ .github/workflows/coverage_report.yml
@@ -8,10 +8,10 @@ jobs:
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v4
-    - name: Set up Python ${{ matrix.version }}
+    - name: Set up Python 3.12
       uses: actions/setup-python@v5
       with:
-        python-version: ${{ matrix.version }}
+        python-version: 3.12
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
@@ -19,12 +19,22 @@ jobs:
         pip install -r dev_requirements.txt
         pip install .
     - name: Test with pytest
-      run: pytest --cov --cov-branch --cov-report=xml;
+      run: pytest --cov --cov-branch --cov-report=xml
+    - name: Upload coverage artifact
+      uses: actions/upload-artifact@v4
+      with:
+        name: coverage-report
+        path: coverage.xml
 
   create-report:
     needs: test
     runs-on: ubuntu-latest
     steps:
+    - uses: actions/checkout@v4
+    - name: Download coverage artifact
+      uses: actions/download-artifact@v4
+      with:
+        name: coverage-report
     - name: Upload coverage reports to Codecov
       uses: codecov/codecov-action@v5
       with:
