+++ .github/workflows/coverage_report.yml
@@ -19,7 +19,7 @@ jobs:
         pip install -r dev_requirements.txt
         pip install .
     - name: Test with pytest
-      run: pytest --cov --cov-branch --cov-report=xml
+      run: pytest --cov=dython --cov-branch --cov-report=xml
     - name: Upload coverage artifact
       uses: actions/upload-artifact@v4
       with:
