+++ .github/workflows/coverage_report.yml
@@ -17,9 +17,9 @@ jobs:
         python -m pip install --upgrade pip
         pip install -r requirements.txt
         pip install -r dev_requirements.txt
-        pip install .
+        pip install -e .
     - name: Test with pytest
-      run: pytest --cov=dython --cov-branch --cov-report=xml
+      run: pytest tests --cov=dython --cov-branch --cov-report=xml
     - name: Upload coverage artifact
       uses: actions/upload-artifact@v4
       with:
