+++ .github/workflows/pythonpublish.yml
@@ -25,9 +25,14 @@ jobs:
           python -m pip install --upgrade pip
           pip install -r requirements.txt
           pip install -r dev_requirements.txt
-          pip install .
+          pip install -e .
       - name: Test with pytest
-        run: pytest
+        run: pytest tests --cov=dython --cov-branch --cov-report=xml
+      - name: Upload coverage artifact
+        uses: actions/upload-artifact@v4
+        with:
+          name: coverage-report
+          path: coverage.xml
 
   deploy:
     needs: test
@@ -54,6 +59,11 @@ jobs:
   #   needs: deploy
   #   runs-on: ubuntu-latest
   #   steps:
+  #   - uses: actions/checkout@v4
+  #   - name: Download coverage artifact
+  #     uses: actions/download-artifact@v4
+  #     with:
+  #       name: coverage-report
   #   - name: Upload coverage reports to Codecov
   #     uses: codecov/codecov-action@v5
   #     with:
