### 2025-11-08 c502caf
Add coverage badge to README
Added a coverage badge to the README.
