### 2025-11-08 4081b98
Fix badge link formatting in README.md
