+++ tests/test_model_utils/test_metric_graph_advanced.py
@@ -1,0 +1,195 @@
+import pytest
+import numpy as np
+import matplotlib.pyplot as plt
+from dython.model_utils import metric_graph
+
+
+def test_metric_graph_invalid_metric():
+    """Test metric_graph with invalid metric"""
+    y_true = [0, 1, 0, 1]
+    y_pred = [0.1, 0.9, 0.3, 0.8]
+    
+    with pytest.raises(ValueError):
+        metric_graph(y_true, y_pred, metric='invalid', plot=False)
+
+
+def test_metric_graph_none_metric():
+    """Test metric_graph with None metric"""
+    y_true = [0, 1, 0, 1]
+    y_pred = [0.1, 0.9, 0.3, 0.8]
+    
+    with pytest.raises(ValueError):
+        metric_graph(y_true, y_pred, metric=None, plot=False)
+
+
+def test_metric_graph_multiclass_with_class_names():
+    """Test metric_graph with multiclass and class names"""
+    y_true = np.array([[1, 0, 0], [0, 1, 0], [0, 0, 1], [1, 0, 0]])
+    y_pred = np.array([[0.7, 0.2, 0.1], [0.1, 0.8, 0.1], [0.2, 0.1, 0.7], [0.6, 0.3, 0.1]])
+    
+    result = metric_graph(y_true, y_pred, metric='roc', class_names=['A', 'B', 'C'], plot=False)
+    assert 'A' in result
+    assert 'B' in result
+    assert 'C' in result
+
+
+def test_metric_graph_multiclass_wrong_class_names_type():
+    """Test metric_graph with multiclass and wrong class names type"""
+    y_true = np.array([[1, 0, 0], [0, 1, 0], [0, 0, 1], [1, 0, 0]])
+    y_pred = np.array([[0.7, 0.2, 0.1], [0.1, 0.8, 0.1], [0.2, 0.1, 0.7], [0.6, 0.3, 0.1]])
+    
+    # class_names as non-list/non-string for multiclass should raise error
+    with pytest.raises((ValueError, TypeError)):
+        metric_graph(y_true, y_pred, metric='roc', class_names=123, plot=False)
+
+
+def test_metric_graph_multiclass_wrong_class_names_count():
+    """Test metric_graph with multiclass and wrong number of class names"""
+    y_true = np.array([[1, 0, 0], [0, 1, 0], [0, 0, 1], [1, 0, 0]])
+    y_pred = np.array([[0.7, 0.2, 0.1], [0.1, 0.8, 0.1], [0.2, 0.1, 0.7], [0.6, 0.3, 0.1]])
+    
+    with pytest.raises(ValueError):
+        metric_graph(y_true, y_pred, metric='roc', class_names=['A', 'B'], plot=False)
+
+
+def test_metric_graph_pr_binary():
+    """Test PR curve for binary classification"""
+    y_true = [0, 1, 0, 1, 1, 0]
+    y_pred = [0.1, 0.9, 0.3, 0.8, 0.7, 0.2]
+    
+    result = metric_graph(y_true, y_pred, metric='pr', plot=False)
+    assert 'auc' in result['0']
+    assert 'naive' in result['0']['auc']
+
+
+def test_metric_graph_pr_multiclass():
+    """Test PR curve for multiclass"""
+    y_true = np.array([[1, 0, 0], [0, 1, 0], [0, 0, 1], [1, 0, 0]])
+    y_pred = np.array([[0.7, 0.2, 0.1], [0.1, 0.8, 0.1], [0.2, 0.1, 0.7], [0.6, 0.3, 0.1]])
+    
+    result = metric_graph(y_true, y_pred, metric='pr', plot=False)
+    assert '0' in result
+
+
[…]
