### 2025-11-10 39850e1
more tests and removing examples
