+++ tests/test_private_helpers_advanced.py
@@ -1,0 +1,277 @@
+import pytest
+import numpy as np
+import pandas as pd
+import matplotlib.pyplot as plt
+from dython._private import (
+    convert,
+    remove_incomplete_samples,
+    replace_nan_with_value,
+    plot_or_not,
+    set_is_jupyter,
+)
+
+
+class TestConvertAdditional:
+    """Additional tests for convert function to increase coverage"""
+    
+    def test_convert_dataframe_to_ndarray(self):
+        """Test converting DataFrame to ndarray (line 69)"""
+        df = pd.DataFrame({'a': [1, 2, 3], 'b': [4, 5, 6]})
+        result = convert(df, np.ndarray)
+        assert isinstance(result, np.ndarray)
+        assert result.shape == (3, 2)
+    
+    def test_convert_dataframe_to_ndarray_no_copy(self):
+        """Test converting DataFrame to ndarray without copy"""
+        df = pd.DataFrame({'a': [1, 2, 3], 'b': [4, 5, 6]})
+        result = convert(df, np.ndarray, copy=False)
+        assert isinstance(result, np.ndarray)
+    
+    def test_convert_series_to_list(self):
+        """Test converting Series to list (line 76)"""
+        series = pd.Series([1, 2, 3, 4, 5])
+        result = convert(series, list)
+        assert isinstance(result, list)
+        assert result == [1, 2, 3, 4, 5]
+    
+    def test_convert_ndarray_to_list(self):
+        """Test converting ndarray to list (line 78)"""
+        arr = np.array([1, 2, 3, 4, 5])
+        result = convert(arr, list)
+        assert isinstance(result, list)
+        assert result == [1, 2, 3, 4, 5]
+    
+    def test_convert_ndarray_to_list_2d(self):
+        """Test converting 2D ndarray to list"""
+        arr = np.array([[1, 2], [3, 4], [5, 6]])
+        result = convert(arr, list)
+        assert isinstance(result, list)
+        assert len(result) == 3
+        assert result[0] == [1, 2]
+    
+    def test_convert_list_to_list_no_copy(self):
+        """Test converting list to list without copy"""
+        lst = [1, 2, 3, 4, 5]
+        result = convert(lst, list, copy=False)
+        assert result is lst  # Should be the same object
+    
+    def test_convert_list_to_list_with_copy(self):
+        """Test converting list to list with copy"""
+        lst = [1, 2, 3, 4, 5]
+        result = convert(lst, list, copy=True)
+        assert result is not lst  # Should be a different object
+        assert result == lst  # But with same values
+    
+    def test_convert_ndarray_to_ndarray_no_copy(self):
+        """Test converting ndarray to ndarray without copy"""
+        arr = np.array([1, 2, 3, 4, 5])
+        result = convert(arr, np.ndarray, copy=False)
+        assert result is arr  # Should be the same object
+    
+    def test_convert_ndarray_to_ndarray_with_copy(self):
+        """Test converting ndarray to ndarray with copy"""
+        arr = np.array([1, 2, 3, 4, 5])
+        result = convert(arr, np.ndarray, copy=True)
+        assert result is not arr  # Should be a different object
+        np.testing.assert_array_equal(result, arr)  # But with same values
+    
+    def test_convert_dataframe_to_dataframe_no_copy(self):
+        """Test converting DataFrame to DataFrame without copy"""
+        df = pd.DataFrame({'a': [1, 2, 3]})
+        result = convert(df, pd.DataFrame, copy=False)
+        assert result is df  # Should be the same object
+    
+    def test_convert_dataframe_to_dataframe_with_copy(self):
[…]
