+++ tests/test_data_utils/test_identify_columns.py
@@ -1,0 +1,242 @@
+import pytest
+import numpy as np
+import pandas as pd
+from dython.data_utils import identify_columns_by_type, identify_columns_with_na
+
+
+class TestIdentifyColumnsByType:
+    """Tests for identify_columns_by_type function"""
+    
+    def test_identify_int_columns(self):
+        """Test identifying integer columns"""
+        df = pd.DataFrame({
+            'int_col': [1, 2, 3, 4],
+            'float_col': [1.0, 2.0, 3.0, 4.0],
+            'str_col': ['a', 'b', 'c', 'd']
+        })
+        
+        result = identify_columns_by_type(df, include=['int64'])
+        assert 'int_col' in result
+        assert 'float_col' not in result
+        assert 'str_col' not in result
+    
+    def test_identify_float_columns(self):
+        """Test identifying float columns"""
+        df = pd.DataFrame({
+            'int_col': [1, 2, 3, 4],
+            'float_col': [1.0, 2.0, 3.0, 4.0],
+            'str_col': ['a', 'b', 'c', 'd']
+        })
+        
+        result = identify_columns_by_type(df, include=['float64'])
+        assert 'float_col' in result
+        assert 'int_col' not in result
+        assert 'str_col' not in result
+    
+    def test_identify_object_columns(self):
+        """Test identifying object columns"""
+        df = pd.DataFrame({
+            'int_col': [1, 2, 3, 4],
+            'float_col': [1.0, 2.0, 3.0, 4.0],
+            'str_col': ['a', 'b', 'c', 'd']
+        })
+        
+        result = identify_columns_by_type(df, include=['object'])
+        assert 'str_col' in result
+        assert 'int_col' not in result
+        assert 'float_col' not in result
+    
+    def test_identify_multiple_types(self):
+        """Test identifying multiple column types"""
+        df = pd.DataFrame({
+            'int_col': [1, 2, 3, 4],
+            'float_col': [1.0, 2.0, 3.0, 4.0],
+            'str_col': ['a', 'b', 'c', 'd']
+        })
+        
+        result = identify_columns_by_type(df, include=['int64', 'float64'])
+        assert 'int_col' in result
+        assert 'float_col' in result
+        assert 'str_col' not in result
+    
+    def test_identify_category_columns(self):
+        """Test identifying categorical columns"""
+        df = pd.DataFrame({
+            'int_col': [1, 2, 3, 4],
+            'cat_col': pd.Categorical(['a', 'b', 'c', 'a'])
+        })
+        
+        result = identify_columns_by_type(df, include=['category'])
+        assert 'cat_col' in result
+        assert 'int_col' not in result
+    
+    def test_identify_with_numpy_array(self):
+        """Test identify_columns_by_type with numpy array"""
+        arr = np.array([[1, 2, 3], [4, 5, 6]])
+        
+        result = identify_columns_by_type(arr, include=['int64'])
+        # Numpy arrays converted to DataFrame have default numeric types
+        assert isinstance(result, list)
+    
+    def test_identify_no_matching_columns(self):
+        """Test when no columns match the requested type"""
+        df = pd.DataFrame({
+            'int_col': [1, 2, 3, 4],
+            'float_col': [1.0, 2.0, 3.0, 4.0]
+        })
+        
+        result = identify_columns_by_type(df, include=['object'])
+        assert result == []
+    
+    def test_identify_all_columns_match(self):
+        """Test when all columns match the requested type"""
+        df = pd.DataFrame({
+            'col1': [1, 2, 3, 4],
+            'col2': [5, 6, 7, 8],
+            'col3': [9, 10, 11, 12]
+        })
+        
[…]
