+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 0.7.12 (dev)
+* Added new tests (issue [#172](https://github.com/shakedzy/dython/issues/172))
+* `examples` module removed (all examples exist in the [official documentation](https://shakedzy.xyz/dython/getting_started/examples/))
+
 ## 0.7.11
 * Fixing dependency issue ([#170](https://github.com/shakedzy/dython/issues/170))
 * Resolving multiple typing issues and warnings
