+++ .github/workflows/pytest.yml
@@ -32,7 +32,7 @@ jobs:
   build:
     strategy:
       matrix:
-        version: ["3.10", "3.13"]
+        version: ["3.10", "3.14"]
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v4
