+++ .github/workflows/pythonpublish.yml
@@ -12,7 +12,7 @@ jobs:
   test:
     strategy:
       matrix:
-        version: [ "3.10", "3.11", "3.12", "3.13" ]
+        version: [ "3.10", "3.11", "3.12", "3.13", "3.14" ]
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
