### 2025-11-10 #9174 @sepandhaghighi
> 1. tests _do_ run across all Python versions. I see now that my workflows didn't transition to the `joss_paper` branch, but you can see it [here](https://github.com/shakedzy/dython/blob/master/.github/workflows/pythonpublish.yml). Before any release is sent to PyPI, all tests need to pass across all Python versions. They currently all run on Linux, but I can add other platforms too. I added these to the `joss_paper` branch now, and commented out their triggers.
> 2. Implemented CodeCov + added tests -> coverage is now 90.73%

@shakedzy

Great job! 💯
I highly recommend that you test the code on both `macOS` and `Windows` too. Additionally, it’s a good idea to separate your regression testing from the PyPI upload workflow. Regression tests should be triggered at least on each pull request, even if you choose not to run them on every commit.

> 3. Not sure I follow the request about the tests - there are dedicated tests for both Cramer's V the Theil's U, both run on well known and used datasets, thus allowing users to review their results. Tests were also peer-reviewed. Am I missing anything?

I see that you are testing `Cramer's V` on the iris dataset in `tests/test_nominal/test_cramers_v.py`. However, what I'm asking about is verified tests. Are there any papers, websites, or tools that compute this parameter for this dataset and yield the result `0.14201914309546954`? If such sources are available, please leave a comment and reference them near each test.

> Regarding Quality -
Not sure I understand how a block diagram would be beneficial, as most methods are independent (unless private). Could you please elaborate on this?

I referred to a block diagram to illustrate the different modules of the package. However, if you believe that a block diagram is not the best approach, consider using a table or another format instead. In my opinion, including multiple code examples in the paper can reduce its readability.

Best Regards
SH
