+++ .github/workflows/test_and_publish.yml
@@ -1,7 +1,7 @@
 # This workflows will upload a Python Package using Twine when a release is created
 # For more information see: https://help.github.com/en/actions/language-and-framework-guides/using-python-with-github-actions#publishing-to-package-registries
 
-name: Upload Python Package
+name: Upload Python Package to PyPI
 
 on:
   workflow_dispatch: 
@@ -55,16 +55,16 @@ jobs:
         python setup.py sdist bdist_wheel
         twine upload dist/*
 
-  # post-deploy:
-  #   needs: deploy
-  #   runs-on: ubuntu-latest
-  #   steps:
-  #   - uses: actions/checkout@v4
-  #   - name: Download coverage artifact
-  #     uses: actions/download-artifact@v4
-  #     with:
-  #       name: coverage-report
-  #   - name: Upload coverage reports to Codecov
-  #     uses: codecov/codecov-action@v5
-  #     with:
-  #       token: ${{ secrets.CODECOV_TOKEN }}
+  post-deploy:
+    needs: deploy
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v4
+    - name: Download coverage artifact
+      uses: actions/download-artifact@v4
+      with:
+        name: coverage-report
+    - name: Upload coverage reports to Codecov
+      uses: codecov/codecov-action@v5
+      with:
+        token: ${{ secrets.CODECOV_TOKEN }}
