### 2025-11-17 0b2cdac
updating tests for multiple platforms
