+++ .github/workflows/test_code.yml
@@ -29,7 +29,7 @@ permissions:
   contents: read
 
 jobs:
-  build:
+  test:
     strategy:
       matrix:
         version: ["3.10", "3.14"]
