### 2025-12-09 #9174 @sepandhaghighi
My final comment:

From the perspective of the JOSS journal checklist, `dython` seems satisfactory. Nevertheless, I think that tools like `dython`, which compute various coefficients, require valid test cases that have been verified against a golden value. This point is not explicitly mentioned in the JOSS rules, so I have left it for the editor to decide about it.

Best Regards
SH
