### 2025-12-11 #9174 @sepandhaghighi
> hey [@sepandhaghighi](https://github.com/sepandhaghighi) - the tests point have been brought up in earlier comments:
>
> > 2. I'm not aware of any papers that present such data. To be honest, academic papers seem to rely on this package for computations (see [here](https://arxiv.org/pdf/2204.00401), [here](https://arxiv.org/pdf/2307.07997) and [here](https://dl.acm.org/doi/pdf/10.1145/3583780.3615202) for example)
>
> Hope that is satisfactory

Everything appears good in order with the multi-platform tests. However, regarding the verified tests, I'm aware that many papers utilize this tool to calculate coefficients such as Cramer's V, but this does not guarantee the correctness of the implementation. I'm familiar with an application that was used by approximately **500** papers to calculate various network parameters, only to discover that the tool was not accurate enough!! This is because authors are unable to independently verify the implementations used by the tools they rely on.

Upon manually reviewing your implementation, it appears to be accurate. However, I strongly recommend adding verified tests to `dython`. If a scientific paper with a golden value cannot be found, it would be better to use output from other tools rather than lacking verified tests. Additionally, some examples of Cramer's V can be found here, and further resources may be available in the R-Language community.

1. http://vassarstats.net/newcs.html
2. https://mathcracker.com/cramers-v-calculator
3. https://web.archive.org/web/20190303120107/https://alias-i.com/lingpipe/docs/api/com/aliasi/classify/ConfusionMatrix.html

Best Regards
SH
