### 2025-12-15 #9174 @shakedzy
@sepandhaghighi while I too would have liked to have the golden value paper, it seems as one cannot be found. As that is the case, comparing to other tools - which probably don't have a golden value to compare too - seems as missing the point of comparing to a golden value.
Combining this with the fact that, as you've seen for yourself, the math is quite straightforward and can be easily verified, I do not see the added value of comparing to other tools.
