### 2025-12-16 d399aaa
Small edits for readability
See https://github.com/openjournals/joss-reviews/issues/9174\#issuecomment-3657982006
