+++ paper.md
@@ -64,7 +64,7 @@ Below is a summary of existing methods of `dython`, per module.
 |--------|-------------|
 | associations | Computes associations between mixed-type features. |
 | cluster_correlations | Applies clustering to reorder a correlation matrix. |
-| compute_associations | Deprecated; replaced by `associations(compute_only)`. |
+| compute_associations | Deprecated; replaced by `associations(compute_only=True)`. |
 | conditional_entropy | Computes conditional entropy of X given Y. |
 | correlation_ratio | Computes correlation between categorical and numeric variables. |
 | cramers_v | Computes Cramér’s V between categorical variables. |
