+++ dython/model_utils.py
@@ -5,7 +5,7 @@ from sklearn.ensemble import RandomForestClassifier
 from sklearn.metrics import roc_curve, precision_recall_curve, auc
 from sklearn.preprocessing import LabelEncoder
 from typing import Any, Iterable
-from .typing import Number, OneDimArray
+from .typing import Number, OneDimArray, MetricGraphResult, SingleCurveResult, SingleMethodResult
 from ._private import convert, plot_or_not
 
 __all__ = ["random_forest_feature_importance", "metric_graph", "ks_abc"]
@@ -53,29 +53,38 @@ def _draw_estimated_optimal_threshold_mark(
     ms: int,
     fmt: str,
     ax: Axes,
-) -> tuple[Number, Number, Number]:
+) -> list[tuple[Number, Number, Number]]:
     annotation_offset = (-0.027, 0.03)
     a = np.zeros((len(x_axis), 2))
     a[:, 0] = x_axis
     a[:, 1] = y_axis
+    a = a[a[:, 0] != a[:, 1]]
     if metric == "roc":
-        dist = lambda row: row[0] ** 2 + (1 - row[1]) ** 2  # optimal: (0,1)
+        dists = [ # optimal: (0,1)
+            lambda row: row[0] ** 2 + (1 - row[1]) ** 2,        # geo
+            lambda row: row[0] - row[1]  # Inverse Youden's J (X-Y instead of Y-X) as later on we're finding the min value, and Youden's J needs to be maximized
+        ]
     else:  # metric == 'pr'
-        dist = (
-            lambda row: (1 - row[0]) ** 2 + (1 - row[1]) ** 2
-        )  # optimal: (1,1)
-    amin = np.apply_along_axis(dist, 1, a).argmin()
-    ax.plot(x_axis[amin], y_axis[amin], color=color, marker="o", ms=ms)     # pyright: ignore[reportCallIssue, reportArgumentType]
-    ax.annotate(
-        "{th:{fmt}}".format(th=thresholds[amin], fmt=fmt),                  # pyright: ignore[reportCallIssue, reportArgumentType]
-        xy=(x_axis[amin], y_axis[amin]),                                    # pyright: ignore[reportCallIssue, reportArgumentType]
-        color=color,
-        xytext=(
-            x_axis[amin] + annotation_offset[0],                            # pyright: ignore[reportCallIssue, reportArgumentType, reportOperatorIssue]
-            y_axis[amin] + annotation_offset[1],                            # pyright: ignore[reportCallIssue, reportArgumentType, reportOperatorIssue]  
-        ),
-    )
-    return thresholds[amin], x_axis[amin], y_axis[amin]                     # pyright: ignore[reportCallIssue, reportArgumentType, reportReturnType]
+        dists = [ # optimal: (1,1)
+            lambda row: (1 - row[0]) ** 2 + (1 - row[1]) ** 2   # geo
+          ] 
+    output_tuples = []
+    for dist, marker in zip(dists, ['o','x']):
+        amin = np.apply_along_axis(dist, 1, a).argmin()
+        ax.plot(x_axis[amin], y_axis[amin], color=color, marker=marker, ms=ms)      # pyright: ignore[reportCallIssue, reportArgumentType]
+        ax.annotate(
+            "{th:{fmt}}".format(th=thresholds[amin], fmt=fmt),                      # pyright: ignore[reportCallIssue, reportArgumentType]
+            xy=(x_axis[amin], y_axis[amin]),                                        # pyright: ignore[reportCallIssue, reportArgumentType]
+            color=color,
+            xytext=(
+                x_axis[amin] + annotation_offset[0],                                # pyright: ignore[reportCallIssue, reportArgumentType, reportOperatorIssue]
[…]
