### 2025-12-16 dea34e8
Update CHANGELOG.md
