### 2025-12-16 6eb704d
Fix syntax error in ROC graph plotting
