### 2025-12-17 fb55e4a
minor changes in paper.md
