+++ paper.md
@@ -31,25 +31,25 @@ In short, **`dython`** lowers the friction for inter-variable association analys
 
 # Statement of Need  
 
-While there are many statistical and visualization libraries in Python (e.g. `pandas` [@pandas], `scipy` [@scipy], `scikit-learn` [@scikit-learn], `seaborn` [@seaborn]), they treat continuous data, categorical data and the overall visualization separately. Users often resort to custom glue code to:  
+While there are many statistical and visualization libraries in Python (e.g., `pandas` [@pandas], `scipy` [@scipy], `scikit-learn` [@scikit-learn], `seaborn` [@seaborn]), they treat continuous data, categorical data, and the overall visualization separately. Users often resort to custom glue code to:  
 
 1. determine which columns are categorical vs. numeric,  
-2. choose an appropriate association statistic (e.g. Pearson's R for numeric–numeric, correlation ratio for numeric–categorical, Cramér’s V or Theil’s U for categorical–categorical),  
+2. choose an appropriate association statistic (e.g., Pearson's R for numeric–numeric, correlation ratio for numeric–categorical, Cramér’s V or Theil’s U for categorical–categorical),  
 3. compute those pairwise,  
-4. assemble a matrix or graph,  
+4. assemble a matrix or graph, and
 5. annotate, visualize, and interpret the results.
 
 This fragmentation results in boilerplate, inconsistency, or risk of mistakes, especially in exploratory settings or pipelines.  
 
 **`dython`** addresses this gap by providing a unified, high-level API that:  
 
-- **infers variable types** 
+- **infers variable types**, 
  
-- **automatically selects appropriate measures** 
+- **automatically selects appropriate measures**,
  
-- **returns structured and annotated output**
+- **returns structured and annotated output**,
   
-- **offers visualization** (heatmaps, annotation) integrations
+- **offers visualization** (heatmaps, annotation) integrations, and
    
 - **offers model evaluation tools** (ROC, AUC, thresholding) for classification tasks
 
@@ -62,47 +62,47 @@ Below is a summary of existing methods of `dython`, per module.
 
 | Method | Description |
 |--------|-------------|
-| associations | Computes associations between mixed-type features. |
-| cluster_correlations | Applies clustering to reorder a correlation matrix. |
-| compute_associations | Deprecated; replaced by `associations(compute_only=True)`. |
-| conditional_entropy | Computes conditional entropy of X given Y. |
-| correlation_ratio | Computes correlation between categorical and numeric variables. |
-| cramers_v | Computes Cramér’s V between categorical variables. |
-| identify_nominal_columns | Detects nominal (categorical) columns. |
-| identify_numeric_columns | Detects numeric columns. |
-| numerical_encoding | Encodes a mixed dataset into numeric format. |
-| replot_last_associations | Replots the last association heatmap. |
-| theils_u | Computes Theil’s U (uncertainty coefficient). |
+| associations | Computes associations between mixed-type features |
+| cluster_correlations | Applies clustering to reorder a correlation matrix |
+| compute_associations | Deprecated; replaced by `associations(compute_only=True)` |
+| conditional_entropy | Computes conditional entropy of X given Y |
+| correlation_ratio | Computes correlation between categorical and numeric variables |
+| cramers_v | Computes Cramér’s V between categorical variables |
+| identify_nominal_columns | Detects nominal (categorical) columns |
+| identify_numeric_columns | Detects numeric columns |
+| numerical_encoding | Encodes a mixed dataset into numeric format |
+| replot_last_associations | Replots the last association heatmap |
+| theils_u | Computes Theil’s U (uncertainty coefficient) |
 
 ## `model_utils`
 
 | Method | Description |
[…]
