### 2025-12-17 #9174 @danielskatz
> * "boilerplate" in line 42 means exactly what you wrote: _"excessive copying and pasting"_. How would you phrase this?

Maybe just that then, or perhaps "excessive use of boilerplate code"?
