### 2025-12-17 f71068d
Update paper.md
