### 2025-12-17 6d1b0af
fix pandas citation
