+++ paper.bib
@@ -22,20 +22,14 @@
 }
 
 @software{pandas,
-  author       = {Pandas development team},
+  author       = {{Pandas development team}},
   title        = {pandas-dev/pandas: {P}andas},
-  month        = sep,
-  year         = 2025,
+  month        = feb,
+  year         = 2020,
   publisher    = {Zenodo},
-  version      = {v2.3.3},
-  doi          = {10.5281/zenodo.17229934},
-  url          = {https://doi.org/10.5281/zenodo.17229934},
-  swhid        = {swh:1:dir:5bc73e541e362cf8e7068ebb434b9d7bc1194c14
-                   ;origin=https://doi.org/10.5281/zenodo.3509134;vis
-                   it=swh:1:snp:8c6e563b81f60a3d62fec309b38f470a0eacf
-                   596;anchor=swh:1:rel:b0f463d37766850a89633bd05643d
-                   45342b9241d;path=pandas-dev-pandas-1efe649
-                  },
+  version      = {latest},
+  doi          = {10.5281/zenodo.3509134},
+  url          = {https://doi.org/10.5281/zenodo.3509134}
 }
 
 @article{scipy,
