### 2025-12-17 cfd47f6
Update CHANGELOG.md
