### 2025-12-17 147ef89
Update model_utils.md
