### 2025-12-17 02745bb
Create CITATION.cff
