### 2025-12-17 12e3333
Add JOSS paper link
