+++ docs/index.md
@@ -9,7 +9,8 @@ is_homepage:
 [![Python Version](https://img.shields.io/pypi/pyversions/dython.svg?style=for-the-badge)](https://pypi.org/project/dython/)
 [![PyPI - Downloads](https://img.shields.io/pypi/dm/dython?style=for-the-badge)](https://pypistats.org/packages/dython)
 [![License](https://img.shields.io/pypi/l/dython?style=for-the-badge)](https://github.com/shakedzy/dython/blob/master/LICENSE)
-[![DOI](https://img.shields.io/badge/DOI-10.5281%2Fzenodo.12698421-eb34c6?style=for-the-badge)](https://zenodo.org/doi/10.5281/zenodo.12698421)
+[![Paper](https://img.shields.io/badge/JOSS-10.21105%2Fjoss.09174-eb34c6?style=for-the-badge)](https://joss.theoj.org/papers/10.21105/joss.09174)
+[![Zenodo](https://img.shields.io/badge/Zenodo-10.5281%2Fzenodo.12698421-8f34eb?style=for-the-badge)](https://zenodo.org/doi/10.5281/zenodo.12698421)
 
 ![banner](images/index_banner.png)
 
@@ -58,13 +59,18 @@ for more information.
 See some usage examples of `nominal.associations` and `model_utils.roc_graph` on the [examples page](getting_started/examples.md).
 
 ## Citing
-Use this reference to cite if you use Dython in a paper:
+When using Dython, please cite it using this citation:
 ```bibtex
-@software{Zychlinski_dython_2018,
-  author = {Zychlinski, Shaked},
-  title = {{dython}},
-  year = {2018},
-  url = {https://github.com/shakedzy/dython},
-  doi = {10.5281/zenodo.12698421}
-}
+@article{Zychlinski2025,
+    doi = {10.21105/joss.09174},
+    url = {https://doi.org/10.21105/joss.09174},
+    year = {2025},
+    publisher = {The Open Journal},
+    volume = {10},
+    number = {116},
+    pages = {9174},
+    author = {Shaked Zychlinski},
+    title = {dython: A Set of Analysis and Visualization Tools for Data and Variables in Python},
+    journal = {Journal of Open Source Software}
+ }
 ```
