### 2026-01-12 7f72a9b
Add GitHub Actions workflow to publish site
